made avgReadCount and avgSendCound to be uint32_t instead of floats, since they count bytes. Changes kb to 1024 instead of 1000

+++ libretroshare/src/pqi/pqistreamer.h
@@ -154,8 +154,8 @@ class pqistreamer: public PQInterface
 		time_t mCurrSentTS;
 
 		time_t mAvgLastUpdate; // TS from which these are measured.
-		float mAvgReadCount;
-		float mAvgSentCount;
+		uint32_t mAvgReadCount;
+		uint32_t mAvgSentCount;
 
 		time_t mLastIncomingTs;
